Discord config split

+++ Platforms/Web/Processing/Api/Discord/Twitchalerts/delete.py
@@ -8,9 +8,11 @@ import discord
 from aiohttp.web import Response, Request
 from Utils.Classes.webrequestcontent import WebRequestContent
 from Utils.Classes.discordwebuserinfo import DiscordWebUserInfo
+from Utils.Classes.discordtwitchalert import DiscordTwitchAlert
+from Platforms.Discord.utils import getDiscordServerTwitchAlerts
 from Platforms.Web.Processing.Api.errors import apiMissingAuthorisation, apiMissingData
 from Platforms.Web.Processing.Api.Discord.errors import apiDiscordGuildUnknown, apiDiscordMemberNotFound, apiDiscordMissingPermission
-from Utils.Classes.undefined import UNDEFINED
+from .errors import apiDiscordAlertNotExists
 
 async def apiDiscordTwitchalertsDelete(cls:"WebIndex", WebRequest:Request) -> Response:
 	"""
@@ -20,8 +22,8 @@ async def apiDiscordTwitchalertsDelete(cls:"WebIndex", WebRequest:Request) -> Re
 	await Data.load()
 
 	# get required vars
-	guild_id:str = Data.getStr("guild_id", UNDEFINED, must_be_digit=True)
-	alert_id:int = Data.getInt("alert_id", UNDEFINED, min_x=1)
+	guild_id:str = Data.getStr("guild_id", "", must_be_digit=True)
+	alert_id:str = Data.getStr("alert_id", "", must_be_digit=1)
 
 	# checks
 	if not guild_id:
@@ -48,17 +50,22 @@ async def apiDiscordTwitchalertsDelete(cls:"WebIndex", WebRequest:Request) -> Re
 	if not (CheckMember.guild_permissions.administrator or CheckMember.guild_permissions.manage_guild):
 		return await apiDiscordMissingPermission(cls, WebRequest, guild_id=guild_id, user_id=DiscordUser.user_id)
 
+	# get alert
+	alert_res:list = await getDiscordServerTwitchAlerts(guild_id, alert_id=alert_id)
+
+	if not alert_res:
+		return await apiDiscordAlertNotExists(cls, WebRequest, alert_id=alert_id)
+
+	AlertToDelete:DiscordTwitchAlert = alert_res.pop(0)
+
 	cls.Web.BASE.PhaazeDB.deleteQuery("""
-		DELETE FROM `discord_twitch_alert`
-		WHERE `discord_twitch_alert`.`discord_guild_id` = %s
-			AND `discord_twitch_alert`.`id` = %s""",
-		(guild_id, alert_id)
+		DELETE FROM `discord_twitch_alert` WHERE `discord_guild_id` = %s AND `id` = %s""",
+		(AlertToDelete.guild_id, AlertToDelete.alert_id)
 	)
 
-	cls.Web.BASE.Logger.debug(f"(API/Discord) Deleted alert: S:{guild_id} A:{alert_id}", require="discord:alert")
-
+	cls.Web.BASE.Logger.debug(f"(API/Discord) Twitchalert: {guild_id=} deleted {alert_id=}", require="discord:alert")
 	return cls.response(
-		text=json.dumps( dict(msg="alert successfull deleted", alert_id=alert_id, status=200) ),
+		text=json.dumps( dict(msg="Twitchalerts: Entry deleted", deleted=AlertToDelete.twitch_channel_name, status=200) ),
 		content_type="application/json",
 		status=200
 	)
